moved value objects to right folders
