+++ src/GameStreamer.Domain/GameStreamer.Domain.csproj
@@ -8,6 +8,8 @@
 
   <ItemGroup>
     <Folder Include="docs\" />
+    <Folder Include="Entities\Invitations\" />
+    <Folder Include="Entities\Roomies\" />
   </ItemGroup>
 
 </Project>
